+++ src/hooks/transactions/helpers/index.ts
@@ -1,1 +1,2 @@
 export * from './getShouldMoveTransactionsToSignedState';
+export * from './useSetTransactionNonces';